refactor(tests): share e2e helpers in tests/common
Move fixture, trellis, git, json_output and the other duplicated helpers
into tests/common/mod.rs; every test file now imports them. Inline
git init/add/commit triples use init_repo and JSON-parsing blocks use
json_output. 413 tests, snapshots unchanged.
